Release Alpha-0.2! (#99)
* Removed all idea files

* Added caddyfile template for docker

* Added Staging support.

* Add buttons to listing pages, Format Logs page (#88)

* Removed all idea files

* Added "Add buttons on listing pages"
Change logs div to pre tag

* Added Staging support.


* New installer page.


* Docker ready for primetime (#93)

Docker should function properly now

* Widened Installer


* Added Caddy Modules


* More script reworking! (#94)

* Docker updated with new caddy curl command

* Removed all sleeps. Replaced with prompts and colored output

Parameterized as much of the directory structure as possible. Will make for easier modification in the future.

* After testing scripts, found some issues. Fixed caddy curl command (needed license type), added a log dir variable, fixed some formatting and flow so output is cleaner when running scripts

* Format issue in uninstall script

* The curl is not working with additional DNS providers. Possibly a license server issue on the part of caddy

* Removed extra docker files
Updated remaining and entrypoint to be capable of building caddy from source due to some issues with curl and their download server

* hotfix for proxy setting changes not regernating caddyfile.conf

* Updated docs

* Reworked the DNS Verification, it now applies to hosts and stores the envs in a file for caddy,

Added some reworked generation code and fixed a few little bugs.


* Ignored docker files in data

* DNS domain verification

DNS domain verification work in progress.

* Added caddyfile template for docker

* Added Staging support.

* Add buttons to listing pages, Format Logs page (#88)

* Removed all idea files

* Added "Add buttons on listing pages"
Change logs div to pre tag

* Release Alpha-0.1a (#92)

Fixed issue #86 

* Update permissions and fix formatting

* Update readme to include chmod step
Update installer to not use echo
Update installer to hopefully self elevate

* Fix formatting

* Mor formatting fixes

* Corrected scripts. Should work now but need to verify with testing

* Missed changes

* readme changes

* Found some permissions and issues calling Caddy for a reload

* small typo

* Fixed odd issue where pidfile was not being read and therefore caddy was not reloading properly

* Hopefully final push of 0.1a

* Final script cleanup for 0.1a

* Added Staging support.


* New installer page.


* Docker ready for primetime (#93)

Docker should function properly now

* Widened Installer


* Added Caddy Modules


* More script reworking! (#94)

* Docker updated with new caddy curl command

* Removed all sleeps. Replaced with prompts and colored output

Parameterized as much of the directory structure as possible. Will make for easier modification in the future.

* After testing scripts, found some issues. Fixed caddy curl command (needed license type), added a log dir variable, fixed some formatting and flow so output is cleaner when running scripts

* Format issue in uninstall script

* The curl is not working with additional DNS providers. Possibly a license server issue on the part of caddy

* Removed extra docker files
Updated remaining and entrypoint to be capable of building caddy from source due to some issues with curl and their download server

* hotfix for proxy setting changes not regernating caddyfile.conf

* Updated docs

* Reworked the DNS Verification, it now applies to hosts and stores the envs in a file for caddy,

Fixed bug where Install does not generate required File.


* Reworked the DNS Verification, it now applies to hosts and stores the envs in a file for caddy,

Fixed bug where Install does not generate required File.


* Reworked the DNS Verification, it now applies to hosts and stores the envs in a file for caddy,

Fixed bug where Install does not generate required File.


* Fixed auth bug.


* Disabled DNS Verification.


* Updated background animation on install and login


* Fixed upgrade script formatting *still untested)
Corrected KNOWN ISSUES since DNS verification isnt available as an option in this release

* Update views.py

Fixed issue with hosts error on creation.

* Fixed caddy service issue
Updated dockerfile

* With TLS off, if HTTP:// not in host def, add it.


* Update README.md

Small readme updates

* Update README.md

Small readme updates.

* Ready for release? (#98)

* No known issues in this release
Added version file for docker
Updated docker build to work with version microbadger
Fixed installer script

Now updater functions properly when upgrading from 0.1a

+++ tyger-uninstall-ubuntu.sh
@@ -1,8 +1,23 @@
 #!/bin/bash
 
+# Set used terminal colors
+RED=$(tput setaf 1)
+BLINK=$(tput blink)
+GREEN=$(tput setaf 2)
+NORMAL=$(tput sgr0)
+LIME_YELLOW=$(tput setaf 190)
+
+# do not change, these are hardcoded elsewhere
+APPS_DIR=/apps
+BACKUP_DIR=/backup
+TYGER_ROOT=$APPS_DIR/TygerCaddy
+TYGER_DIR=$TYGER_ROOT/TygerCaddy
+TYGER_DATA=$TYGER_DIR/data
+TYGER_LOGS=$TYGER_DATA/logs
+TYGER_BACKUP=$BACKUP_DIR/TygerCaddy
+
 if [ "$(whoami)" != 'root' ]; then
-  printf "This script must be run as "root".\n"
-  printf "Enter password to elevate privileges:"
+  printf "${GREEN}This script must be run as \"root\"${NORMAL}\n"
   printf "\n"
   SCRIPTPATH=$( cd $(dirname $0) ; pwd -P )
   SELF=`basename $0`
@@ -10,43 +25,44 @@ if [ "$(whoami)" != 'root' ]; then
   exit 1
 fi
 
-printf "Starting uninstaller\n" \
-     "You have 5 before uninstallation begins...\n" \
-     "or hit Ctrl+C to quit\n" \
-     "\n"
-sleep 6
+printf "${RED}You are about to uninstall TygerCaddy${NORMAL}\n"
+read -p "${LIME_YELLOW}Proceed? (y/N)${NORMAL} " -r
+printf "\n"
+if [[ ! $REPLY =~ ^[Yy*] ]]
+then
+    exit 1
+fi
 
-printf "Stopping Services"
-
+printf "${GREEN}Stopping and removing services...${NORMAL}\n"
 service uwsgi stop
 service caddy stop
 
-printf "Removing Python Packages..."
-
-pip3 uninstall -y -r /apps/TygerCaddy/TygerCaddy/requirements.txt
-
-printf "Removing TygerCaddy\n" \
-     "WARNING! This will remove the whole /apps directory and all backups.\n" \
-     "This is your last chance to cancel\n" \
-     "You have 5 seconds to proceed...\n" \
-     "or hit Ctrl+C to quit\n" \
-     "\n"
-sleep 6
-
 rm -rf /usr/local/bin/caddy \
-      /etc/caddy \
-      /etc/ssl/caddy \
-      /etc/systemd/system/caddy.service \
-      /etc/systemd/system/caddy-reload.path \
-      /etc/systemd/system/caddy-reload.service \
-      /etc/systemd/system/uwsgi.service \
-      /apps/TygerCaddy \
-      /backup/TygerCaddy
+       /etc/systemd/system/caddy.service \
+       /etc/systemd/system/caddy-reload.path \
+       /etc/systemd/system/caddy-reload.service \
+       /etc/systemd/system/uwsgi.service
 
 apt remove -y uwsgi
 systemctl daemon-reload
 
+printf "${GREEN}Remove Python packages?${NORMAL} "
+read -p "${LIME_YELLOW}(y/N)${NORMAL} " -r
+printf "\n"
+if [[ $REPLY =~ ^[Yy*] ]]
+  then
+    pip3 uninstall -y -r $TYGER_DIR/requirements.txt
+fi
 
-printf "Uninstall Complete!"
+printf "${RED}${BLINK}About to uninstall all user data${NORMAL}\n"
+read -p "${RED}Proceed? (y/N)${NORMAL} " -r
+printf "\n"
+if [[ $REPLY =~ ^[Yy*] ]]
+  then
+    rm -rf /etc/caddy \
+           /etc/ssl/caddy \
+           $TYGER_ROOT \
+           $TYGER_BACKUP
+fi
 
-sleep 3
+printf "${GREEN}Uninstall Complete!${NORMAL}\n"
